requirement.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
fastapi
uvicorn
sse_starlette


